Develop prepared statements and transactions for simple test case

+++ app/src/sqlite/java/com/study/benchmarkorm/db/CursorWrappers.java
@@ -22,7 +22,7 @@ public class CursorWrappers {
             String author = getString(getColumnIndex(BookTable.Cols.AUTHOR));
             int pages = getInt(getColumnIndex(BookTable.Cols.PAGES_COUNT));
             int bookId = getInt(getColumnIndex(BookTable.Cols.BOOK_ID));
-            int libId = getInt(getColumnIndex(BookTable.Cols.LIBRARY_ID));
+            long libId = getInt(getColumnIndex(BookTable.Cols.LIBRARY_ID));
             Book book = new Book();
             book.setTitle(title);
             book.setAuthor(author);
@@ -62,7 +62,7 @@ public class CursorWrappers {
             Date date = new Date(getString(getColumnIndex(PersonTable.Cols.DATE)));
             String gender = getString(getColumnIndex(PersonTable.Cols.GENDER));
             long phone = getLong(getColumnIndex(PersonTable.Cols.PHONE));
-            int libId = getInt(getColumnIndex(PersonTable.Cols.LIBRARY_ID));
+            long libId = getInt(getColumnIndex(PersonTable.Cols.LIBRARY_ID));
             Person person = new Person();
             person.setFirstName(fName);
             person.setSecondName(sName);
